### 2023-01-27 #5108 @editorialbot
[PRE REVIEW]: PDF Entity Annotation Tool (PEAT)
**Submitting author:** @chrstahl (<a href="http://orcid.org/0000-0002-2070-1555">Christopher Stahl</a>)
**Repository:** https://github.com/USEPA/peat
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.1
