### 2023-02-05 #5108 @arfon
@chrstahl – could you help me understand which parts of the codebase are novel (i.e., written by you) and which are the main JavaScript dependencies?
