### 2023-02-07 #5108 @chrstahl
@arfon The functionality was built off a previous JS tool so a lot of the code is pretty embedded. Things like Electron integration are 100% new such as _electron-starter.js,_ _electron-wait-react.js_, etc. Features like _Sidebar.js_ and the main _App.js_ contain a large portion of the new code. The component classes were also modified, though not as greatly; _PdfHighlighter.js_ being the most changed.
