### 2023-02-19 #5108 @arfon
Thanks @chrstahl. Did you also author the 'previous JS tool' too? If so, have any papers been authored about that? Based on your inputs thus far, I'm struggling to see if this submission meets our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) test.
