### 2023-02-20 #5108 @chrstahl
I did not author the previous JS tool, which is cited in the paper.  I feel the scholarly effort is well defined in the text, we built off a previous application, the previous application provided the baseline for highlighting a PDF in a web browser, but the data generated could not be saved/exported/or used in anyway.

We built a tool that allows for easy annotation of PDF documents for the purpose of ML dataset creation.  Unlike traditional tools, we did not just convert the document to text then write a wrapper that allows you to annotate that (which ignores all the problems of converting from PDF to Text).  Or traditional PDF annotators that just allow you to draw on the PDF.  Our application allows you to create custom tags for easy labelling, provides functions for auto labeling, importing/exporting functionality, generates annotations textual annotations along with coordinate locations within the pdf, ability to rotate pages to handle vertical annotations, etc.  On top of that the application was wrapped into the Electron framework so it can be used cross-platform and without installation.

I am unaware of any open source software that provides this functionality, the closes would be brat rapid annotation tool (cited in the paper); which accomplishes its task by converting to text like mentioned above.
