### 2023-03-08 #5108 @atrisovic
Hi @chrstahl 👋, thank you for submitting the paper! I will be your editor.
As a first step, I'd like to confirm if you had a look at [the list of potential JOSS reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822), and if you have suggestions for who would be a good fit to review the paper? You may also suggest researchers in the area outside of the list.
