### 2023-03-16 #5108 @bioedd
Hi Ana,

Thank you for your invitation. I can and I want to review this paper.
Please let me know the next steps.

Kind regards,
Eduardo Álvarez
Editor Asociado - Economía Creativa

Av. Constituyentes 455, Col. América
[55] 2789 9000 | centro.edu.mx <http://www3.centro.edu.mx/> | @centro_u
<http://www.instagram.com/centro_u>

El mar, 14 mar 2023 a las 9:25, Ana Trisovic ***@***.***>)
escribió:

> Hi @puruckertom <https://github.com/puruckertom>, @EDiLD2
> <https://github.com/EDiLD2> and @bioedd <https://github.com/bioedd>,
> I am writing to invite you to review a paper titled "PDF Entity Annotation
> Tool (PEAT)", which is currently in the review process at the Journal of
> Open Source Software (JOSS). Your insights and feedback on this work would
> be valuable and much appreciated. To help with the review, we have a
> reviewers' checklist
> <https://joss.readthedocs.io/en/latest/review_checklist.html#review-checklist>
> with useful guidelines.
> Thank you!
> Ana Trisovic, JOSS Topic Editor
>
