### 2023-03-18 #5108 @bioedd
Thank you so much.

I will waiting your instructions.

Have a nice day.

Eduardo Álvarez
Editor Asociado - Economía Creativa

Av. Constituyentes 455, Col. América
[55] 2789 9000 | centro.edu.mx <http://www3.centro.edu.mx/> | @centro_u
<http://www.instagram.com/centro_u>

El vie, 17 mar 2023 a las 7:58, Ana Trisovic ***@***.***>)
escribió:

> Hi @bioedd <https://github.com/bioedd>, that's great to hear. Many
> thanks! I will assign you as a reviewer, but we first need to identify one
> more reviewer before the review process starts.
>
