### 2023-03-30 #5108 @puruckertom
Yes, I would be happy to review the paper.
best,
Tom

On Thu, Mar 30, 2023 at 9:36 AM Ana Trisovic ***@***.***>
wrote:

> Hi @puruckertom <https://github.com/puruckertom>, @EDiLD2
> <https://github.com/EDiLD2>,
> This is a kind reminder about my previous message - would you be able to
> review this paper?
> Thank you!
> Ana
>
