### 2023-04-12 #5336 @bioedd
Eduardo Álvarez
Editor Asociado - Economía Creativa

Av. Constituyentes 455, Col. América
[55] 2789 9000 | centro.edu.mx <http://www3.centro.edu.mx/> | @centro_u
<http://www.instagram.com/centro_u>

El mar, 4 abr 2023 a las 10:07, Ana Trisovic ***@***.***>)
escribió:

> 👋🏼 @chrstahl <https://github.com/chrstahl> @bioedd
> <https://github.com/bioedd>, @puruckertom <https://github.com/puruckertom>
> this is the review thread for the paper. All of our communications will
> happen here from now on.
>
> As a reviewer, the first step is to create a checklist for your review by
> entering
>
> @editorialbot generate my checklist
>
> as the top of a new comment in this thread.
>
> These checklists contain the JOSS requirements. As you go over the
> submission, please check any items that you feel have been satisfied. The
> first comment in this thread also contains links to the JOSS reviewer
> guidelines.
>
> The JOSS review is different from most other journals. Our goal is to work
> with the authors to help them meet our criteria instead of merely passing
> judgment on the submission. As such, the reviewers are encouraged to submit
> issues and pull requests on the software repository. When doing so, please
> mention openjournals/joss-reviews#5336 so that a link is created to this
> thread (and I can keep an eye on what is happening). Please also feel free
> to comment and ask questions on this thread.
>
> We aim for reviews to be completed within about 2-4 weeks. Please let me
> know if any of you require some more time. We can also use EditorialBot
> (our bot) to set automatic reminders if you know you'll be away for a known
> period of time.
>
> Please feel free to ping me ***@***.*** <https://github.com/atrisovic>)
> if you have any questions/concerns.
>
