### 2023-09-15 #5336 @atrisovic
Hi @bioedd & @puruckertom,
I hope you had a nice summer. I'm writing to see how is the review going and if you need help with anything. I suggest running `@editorialbot generate my checklist` directly in the issue on GitHub as the formatting for the markdown checklist is not supported from email. Thank you!
