### 2023-11-06 #5336 @gasparl
Hi, thanks for the invitation! I'd accept it (though it's not really my topic), but I only have Linux with me, and the current PEAT doesn't seem to have ready Linux builds. If you can prepare that, let me know, otherwise I'll pass, sorry.
