### 2023-11-14 #5336 @flor14
Hello
I can be a reviewer but currently, I am an ARM Mac user. Even if it would be possible to run the software on my computer it would consume me extra time to make it work. I wonder if I will not be a more useful reviewer for another project if this one is Windows only.
