### 2023-11-14 #5336 @chrstahl
It works on Linux and Mac (ARM or non ARM).  Just have to build using node/yarn, can't put the executable on git without paying for apple to sign the application so it won't work to share a prebuilt.
