### 2023-11-22 #5336 @flor14
@chrstahl, I have been checking your software PEAT. Thank you for your submission.
I have written notes below some of the checklist points, I hope you can find my review useful.

## Review checklist - [@flor14 ](https://github.com/flor14)

**Comment:** Yes, there is an MIT License but I wonder if there is a reason why the year and author are not included on it. The first line is not there: https://choosealicense.com/licenses/mit/

- [X] **Contribution and authorship:** Has the submitting author ([@chrstahl](https://github.com/chrstahl)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

**This item doesn't apply** - [ ] **Data sharing:** If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
**This item doesn't apply**- [ ] **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
**This item doesn't apply** - [ ] **Human and animal research:** If the paper contains original data research on human subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.

**Notes:**
I have to make some changes to install and run the software.
1 - When I try to run it I receive some messages about breaking changes. As the app was not working, I ran `npm audit fix`. These packages were updated:

```
    "foreman": "^3.0.1",
    "jest-puppeteer": "^9.0.1",
    "electron": "^27.1.0",
```

There have been multiple updates for `Electron` since 2021. The version you used (11.0.1) was the first one that supported ARM Mac computers, so that was not an issue as you said before.
2. I had to add `contextIsolation: false` in `public/electron.js` to be able to run it. This could be linked to updates in the version of `Electron`, but I would check if this is the appropriate fix to leave the app publically available. This was the error: https://stackoverflow.com/questions/56091343/typeerror-window-require-is-not-a-function

After 1 and 2, the app worked on my Mac M2 computer.

**Notes:** The app works as expected, I am only getting an error when I try to click the annotate button next to  "Choose an entity". This is the error:

- [ ] **Installation instructions:** Is there a clearly stated list of dependencies? Ideally, these should be handled with an automated package management solution.
**Note:** In the repo is described that the app can be used with Linux and Mac but it only exemplifies its use with Windows. If it is not possible to provide PEAT executable versions for Linux and Mac it would be helpful to guide more to the users about what they should do if they are on Windows and what they should do if they are on Mac/Linux. The `README.md` now lists the commands without more context.

My way of thinking about the level of detail of the documentation is keeping in mind who are your users and the potential level of knowledge you expect them to have. To use an annotation tool such as PEAT, should they know `node` and `yarn`? If the answer is no it would be beneficial to be a bit more verbose about what they should do to make PEAT run. In general, describing steps can help, for example, 1- clone this repository, 2-  Download `Node` and `Yarn` (links), .3 - ...

Also, a comment about the repository: there is no title in the `README.md`. I would add a title and place the instructions about how to install the app under the title "Installation instructions" or similar. I think that can help with the clarity of the documentation.

**Note:** I haven't found community guidelines.

**Note:** The summary is clear, but also is almost as long as the rest of the manuscript. I wonder if there would be benefits of reducing the summary and redistributing some of the information in other sections. However, nothing is said formally about a limit on the number of words for the summary, but I think it is logical to expect the summary to be a bit shorter. I think this is up to you.
