### 2024-02-12 #5336 @atrisovic
Hi @RuneBlaze,
I am writing to kindly invite you to review a paper titled "PDF Entity Annotation Tool (PEAT)", which is currently in the review process at the Journal of Open Source Software (JOSS). Your insights and feedback on this work would be valuable and much appreciated. To help with the review, we have [a reviewers' checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#review-checklist) with useful guidelines.
Thank you!
Ana Trisovic, JOSS Topic Editor
