### 2024-02-12 8681f09
Updated dev depencies and expanding readme for clarity.
