### 2024-02-24 USEPA/peat#2 @RuneBlaze
README.md: More Descriptive README
**Edit: Pending merge into this repo.**

Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/5336#issuecomment-1962727864).

The README currently has `NodeJS and Yarn Requirement` as the top level heading. My suggestions are, in order of priority:

 1. ~~Fix the heading to be PEAT~~
 2. ~~Have a high level paragraph describing what PEAT is for general audiences, and potential applications for more specialized audiences~~
 3. ~~Have a ToC pointing to end-user installation instructions and local developmental instructions.~~
 4. ~~Clarify the supported platforms (is it windows for off-the-shelf installation only? Does this support other platform supporting electron things)~~
 5. Personal suggestion: list a concrete application of PEAT (the annotation is for downstream consumption -- which concrete downstream consumption? What have people done with PEAT that is useful using Spacy and PyTorch? Is there a publication that has used PEAT in its process to create data?)
 6. The README is quite long -- having a ToC will help, especially there are lots of images that take up spaces! I was unsure if you have a contributor guide until scrolling all the way down.

These are my personal suggestions (I am just putting these as potential action items to be constructive :) ). It's a nice tool to say the least).
