### 2024-02-24 USEPA/peat#3 @RuneBlaze
Code Cleanup: Remove Comments & Reformat the Code using Formatters
**Edit: Pending merge of the personal branch into this repo.**

Related to [JOSS review](https://github.com/openjournals/joss-reviews/issues/5336#issuecomment-1962727864).

There are a lot of commented out code in the repo, and I think the codebase for longevity and for inviting contributions can benefit from a standardized formatting tool such as https://prettier.io/.

Suggested and potential ways to improve:

 1. Remove comments that are unused code scattered around (no need to be thorough -- but I see quite a bit here and there)
 2. Run `prettier` (or code formatter of your choice) on all the `.js` files.
