### 2024-02-24 USEPA/peat#4 @RuneBlaze
`yarn start`: TypeError [ERR_UNKNOWN_SIGNAL]: Unknown signal: null
Created a fresh install of the repo at `5f54e0c6a16452d022dfd3951e6ce4bb379de573`.

## Error log

```
E:\PEAT\peat>yarn
yarn install v1.19.1
info No lockfile found.
[1/4] Resolving packages...
⠁ (node:13884) [DEP0040] DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.
(Use `node --trace-deprecation ...` to show where the warning was created)
warning @testing-library/jest-dom > css > urix@0.1.0: Please see https://github.com/lydell/urix#deprecated
warning @testing-library/jest-dom > css > source-map-resolve@0.5.3: See https://github.com/lydell/source-map-resolve#deprecated
warning @testing-library/jest-dom > css > source-map-resolve > urix@0.1.0: Please see https://github.com/lydell/urix#deprecated
warning @testing-library/jest-dom > css > source-map-resolve > resolve-url@0.2.1: https://github.com/lydell/resolve-url#deprecated
warning @testing-library/jest-dom > css > source-map-resolve > source-map-url@0.4.1: See https://github.com/lydell/source-map-url#deprecated
warning @testing-library/react > @types/testing-library__react > @types/testing-library__dom@7.5.0: This is a stub types definition. testing-library__dom provides its own type definitions, so you do not need this installed.
warning http2@3.3.7: Use the built-in module in node 9.0.0 or newer, instead
warning popper > rijs.resdir > chokidar@1.7.0: Chokidar 2 will break on node v14+. Upgrade to chokidar 3 with 15x less dependencies.
warning popper > rijs.resdir > chokidar > fsevents@1.2.13: The v1 package contains DANGEROUS / INSECURE binaries. Upgrade to safe fsevents v2
warning popper > wd > request@2.88.0: request has been deprecated, see https://github.com/request/request/issues/3142
warning popper > wd > request > uuid@3.4.0: Please upgrade  to version 7 or higher.  Older versions may use Math.random() in certain circumstances, which is known to be problematic.  See https://v8.dev/blog/math-random for details.
warning popper > wd > request > har-validator@5.1.5: this library is no longer supported
warning popper > rijs.sync > xrs > uws@9.148.0: New code is available at github.com/uNetworking/uWebSockets.js
warning popper > rijs.resdir > chokidar > readdirp > micromatch > snapdragon > source-map-resolve@0.5.3: See https://github.com/lydell/source-map-resolve#deprecated
warning popper > rijs.sync > buble > acorn5-object-spread@4.0.0: acorn>=5.4.1 supports object-spread
warning popper.js@1.16.1: You can find the new Popper v2 at @popperjs/core, this package is dedicated to the legacy v1
warning Resolution field "autoprefixer@10.4.5" is incompatible with requested version "autoprefixer@^10.4.13"
warning react-scripts > workbox-webpack-plugin@6.6.1: this package has been deprecated
warning react-scripts > babel-preset-react-app > @babel/plugin-proposal-class-properties@7.18.6: This proposal has been merged to the ECMAScript standard and thus this plugin is no longer maintained. Please use @babel/plugin-transform-class-properties instead.
warning react-scripts > babel-preset-react-app > @babel/plugin-proposal-nullish-coalescing-operator@7.18.6: This proposal has been merged to the ECMAScript standard and thus this plugin is no longer maintained. Please use @babel/plugin-transform-nullish-coalescing-operator instead.
warning react-scripts > babel-preset-react-app > @babel/plugin-proposal-numeric-separator@7.18.6: This proposal has been merged to the ECMAScript standard and thus this plugin is no longer maintained. Please use @babel/plugin-transform-numeric-separator instead.
warning react-scripts > babel-preset-react-app > @babel/plugin-proposal-private-methods@7.18.6: This proposal has been merged to the ECMAScript standard and thus this plugin is no longer maintained. Please use @babel/plugin-transform-private-methods instead.
warning react-scripts > babel-preset-react-app > @babel/plugin-proposal-optional-chaining@7.21.0: This proposal has been merged to the ECMAScript standard and thus this plugin is no longer maintained. Please use @babel/plugin-transform-optional-chaining instead.
warning react-scripts > source-map-loader > abab@2.0.6: Use your platform's native atob() and btoa() methods instead
warning react-scripts > workbox-webpack-plugin > workbox-build@6.6.1: this package has been deprecated
warning react-scripts > @svgr/webpack > @svgr/plugin-svgo > svgo@1.3.2: This SVGO version is no longer supported. Upgrade to v2.x.x.
warning react-scripts > react-dev-utils > fork-ts-checker-webpack-plugin > memfs@3.6.0: this will be v4
warning react-scripts > webpack-dev-server > webpack-dev-middleware > memfs@3.6.0: this will be v4
warning react-scripts > workbox-webpack-plugin > workbox-build > rollup-plugin-terser@7.0.2: This package has been deprecated and is no longer maintained. Please use @rollup/plugin-terser
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-broadcast-update@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-background-sync@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-broadcast-update > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-background-sync > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-expiration@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-expiration > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-google-analytics@6.6.1: It is not compatible with newer versions of GA starting with v4, as long as you are using GAv3 it should be ok, but the package is not longer being maintained
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-google-analytics > workbox-background-sync@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-google-analytics > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-range-requests@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-range-requests > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-recipes@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-recipes > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-recipes > workbox-expiration@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-navigation-preload@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-navigation-preload > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-strategies@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-google-analytics > workbox-strategies@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-recipes > workbox-strategies@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-strategies > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-routing@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-google-analytics > workbox-routing@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-recipes > workbox-routing@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-routing > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-window@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-window > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-sw@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-cacheable-response@6.6.1: workbox-background-sync@6.6.1
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-recipes > workbox-cacheable-response@6.6.1: workbox-background-sync@6.6.1
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-cacheable-response > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-precaching@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-recipes > workbox-precaching@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-precaching > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-precaching > workbox-routing@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-precaching > workbox-strategies@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-streams@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-streams > workbox-core@6.6.1: this package has been deprecated
warning react-scripts > workbox-webpack-plugin > workbox-build > workbox-streams > workbox-routing@6.6.1: this package has been deprecated
warning react-scripts > @svgr/webpack > @svgr/plugin-svgo > svgo > stable@0.1.8: Modern JS already guarantees Array#sort() is a stable sort, so this library is deprecated. See the compatibility table on MDN: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort#browser_compatibility
warning react-scripts > css-minimizer-webpack-plugin > cssnano > cssnano-preset-default > postcss-svgo > svgo > stable@0.1.8: Modern JS already guarantees Array#sort() is a stable sort, so this library is deprecated. See the compatibility table on MDN: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort#browser_compatibility
warning react-scripts > workbox-webpack-plugin > workbox-build > @rollup/plugin-replace > magic-string > sourcemap-codec@1.4.8: Please use @jridgewell/sourcemap-codec instead
warning react-scripts > jest > jest-cli > jest-config > jest-environment-jsdom > jsdom > abab@2.0.6: Use your platform's native atob() and btoa() methods instead
warning react-scripts > jest > jest-cli > jest-config > jest-environment-jsdom > jsdom > domexception@2.0.1: Use your platform's native DOMException instead
warning react-scripts > jest > jest-cli > jest-config > jest-environment-jsdom > jsdom > data-urls > abab@2.0.6: Use your platform's native atob() and btoa() methods instead
warning react-scripts > jest > jest-cli > jest-config > jest-environment-jsdom > jsdom > w3c-hr-time@1.0.2: Use your platform's native performance.now() and performance.timeOrigin.
warning url-search-params@1.1.0: now available as @ungap/url-search-params
warning electron-builder > app-builder-lib > electron-osx-sign@0.6.0: Please use @electron/osx-sign moving forward. Be aware the API is slightly different
warning electron-builder > app-builder-lib > @electron/universal > asar@3.2.0: Please use @electron/asar moving forward.  There is no API change, just a package name change
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > core-js@2.6.12: core-js@<3.23.3 is no longer maintained and not recommended for usage due to the number of issues. Because of the V8 engine whims, feature detection in old core-js versions could cause a slowdown up to 100x even if nothing is polyfilled. Some versions have web compatibility issues. Please, upgrade your dependencies to the actual version of core-js.
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > request@2.88.2: request has been deprecated, see https://github.com/request/request/issues/3142
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > request > har-validator@5.1.5: this library is no longer supported
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > request > uuid@3.4.0: Please upgrade  to version 7 or higher.  Older versions may use Math.random() in certain circumstances, which is known to be problematic.  See https://v8.dev/blog/math-random for details.
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > @hapi/joi@15.1.1: Switch to 'npm install joi'
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > @hapi/joi > @hapi/topo@3.1.6: This version has been deprecated and is no longer supported or maintained
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > @hapi/joi > @hapi/hoek@8.5.1: This version has been deprecated and is no longer supported or maintained
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > @hapi/joi > @hapi/topo > @hapi/hoek@8.5.1: This version has been deprecated and is no longer supported or maintained
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > @hapi/joi > @hapi/address@2.1.4: Moved to 'npm install @sideway/address'
warning jest-puppeteer > jest-environment-puppeteer > jest-dev-server > wait-on > @hapi/joi > @hapi/bourne@1.3.2: This version has been deprecated and is no longer supported or maintained
warning puppeteer@2.1.1: < 21.5.0 is no longer supported
[2/4] Fetching packages...
info fsevents@2.3.3: The platform "win32" is incompatible with this module.
info "fsevents@2.3.3" is an optional dependency and failed compatibility check. Excluding it from installation.
info fsevents@1.2.13: The platform "win32" is incompatible with this module.
info "fsevents@1.2.13" is an optional dependency and failed compatibility check. Excluding it from installation.
info dmg-license@1.0.11: The platform "win32" is incompatible with this module.
info "dmg-license@1.0.11" is an optional dependency and failed compatibility check. Excluding it from installation.
info iconv-corefoundation@1.1.7: The platform "win32" is incompatible with this module.
info "iconv-corefoundation@1.1.7" is an optional dependency and failed compatibility check. Excluding it from installation.
[3/4] Linking dependencies...
warning " > autoprefixer@10.4.5" has unmet peer dependency "postcss@^8.1.0".
warning " > @testing-library/user-event@7.2.1" has unmet peer dependency "@testing-library/dom@>=5".
warning " > pdfjs-dist@2.2.228" has unmet peer dependency "webpack@^3.0.0 || ^4.0.0-alpha.0 || ^4.0.0".
warning "pdfjs-dist > worker-loader@2.0.0" has unmet peer dependency "webpack@^3.0.0 || ^4.0.0-alpha.0 || ^4.0.0".
warning "react-scripts > eslint-config-react-app > eslint-plugin-flowtype@8.0.3" has unmet peer dependency "@babel/plugin-syntax-flow@^7.14.5".
warning "react-scripts > eslint-config-react-app > eslint-plugin-flowtype@8.0.3" has unmet peer dependency "@babel/plugin-transform-react-jsx@^7.14.9".
warning " > webpack-dev-middleware@5.3.3" has unmet peer dependency "webpack@^4.0.0 || ^5.0.0".
[4/4] Building fresh packages...
success Saved lockfile.
$ electron-builder install-app-deps
  • electron-builder  version=23.6.0
  • loaded configuration  file=package.json ("build" field)
  • loaded parent configuration  preset=react-cra
  • rebuilding native dependencies  dependencies=uws@9.148.0 platform=win32 arch=x64
  • rebuilding native dependency  name=uws version=9.148.0
Done in 124.95s.

E:\PEAT\peat>yarn start
yarn run v1.22.21
$ nf start -p 3000
[WARN] No ENV file found
4:16:10 PM react.1    |  > PEAT@1.0.1 react-start
4:16:10 PM react.1    |  > BROWSER=none react-scripts start
4:16:10 PM electron.1 |  > PEAT@1.0.1 electron-start
4:16:10 PM electron.1 |  > node src/electron-wait-react main.js
4:16:10 PM react.1    |  'BROWSER' is not recognized as an internal or external command,
4:16:10 PM react.1    |  operable program or batch file.
[DONE] Killing all processes with signal  null
node:internal/util:331
  throw new ERR_UNKNOWN_SIGNAL(signal);
  ^

TypeError [ERR_UNKNOWN_SIGNAL]: Unknown signal: null
    at convertToValidSignal (node:internal/util:331:9)
    at ChildProcess.kill (node:internal/child_process:496:5)
    at EventEmitter.<anonymous> (E:\PEAT\peat\node_modules\foreman\lib\proc.js:54:11)
    at EventEmitter.emit (node:events:531:35)
    at ChildProcess.<anonymous> (E:\PEAT\peat\node_modules\foreman\lib\proc.js:50:13)
    at ChildProcess.emit (node:events:519:28)
    at ChildProcess._handle.onexit (node:internal/child_process:294:12) {
  code: 'ERR_UNKNOWN_SIGNAL'
}

Node.js v21.6.2
error Command failed with exit code 1.
info Visit https://yarnpkg.com/en/docs/cli/run for documentation about this command.
```

## Environment

I was using my Windows 10 machine and with Node.js version v21.6.2, and Yarn version `1.22.21`.
