### 2024-02-24 #5336 @RuneBlaze
@chrstahl I just completed my first round of reviews. First of all, having a software that interacts with PDF files can be a pain at times, so congratulations on achieving this technical mark here. I created two issues in the repo that I think will be great for inviting usage and contribution as an open source software:

 1. https://github.com/USEPA/peat/issues/3
 2. https://github.com/USEPA/peat/issues/2

One minor writing suggestion:

> While text mining approaches – including Deep Learning (DL), Artificial Intelligence (AI), and Machine Learning (ML) - continue to expand at a rapid pace, the tools used by researchers to create the labeled datasets required for training, modeling, and evaluation remain rudimentary.

DL is usually a specific kind of ML, so rephrasing will be appreciated.

Other issues:
 1. https://github.com/USEPA/peat/issues/4 (`yarn start` did not work on my setup: pointers will be appreciated)
 2. Adding the instructions for running tests will be appreciated (I think `yarn test` might be it; but was not sure)
