### 2024-02-26 USEPA/peat#4 @chrstahl
Problem with Windows env variables vs Mac/Linux, should be fixed by using cross-env in latest commit in:

#1
