### 2024-02-26 #5336 @chrstahl
@RuneBlaze Looking through your list but the solved the biggest issue that stuck out to me, yarn start failing on windows.

I have a running pull request going, but currently I can't personally push to the EPA's git so have to wait for approvals.  For working through these issues quicker it might be worthwhile to just look at my personal fork at https://github.com/chrstahl/peat .  I have made some changes to the Readme already in my fork, but will continue going through your comments.
