### 2024-02-29 ab1c6bd
REAME rewrite, code cleanup and beautifier.
