### 2024-02-29 USEPA/peat#3 @chrstahl
#1 Cleaned up unused code and ran everything through a beautifier.
