### 2024-03-04 a747d30
Added prettier format script to package.json
