+++ src/components/Tip.js
@@ -40,7 +40,7 @@ class Tip extends Component<Props, State> {
       optionList.add(new Option("Select Type: ", ""));
 
       this.props.tags.annotation_types.forEach((option) =>
-        optionList.add(new Option(option.name, option.id))
+        optionList.add(new Option(option.name, option.id)),
       );
     }
   }
