### 2024-03-13 #5336 @chrstahl
@RuneBlaze I changed around the Readme, not exactly a TOC but I think it fixes the problem of not being to easily see the different sections.  Also cleaned up all the comments/prettified everything based on your comments.  Looks like they did approve the pull but didn't actually merge it so reaching back out.

For testing, there is not currently any automatic testing.  The user guide explains the step by step how to test functionality though.

I'm going to add a definition for the output json tomorrow.  I do agree with your comments that showing some publications of the tool would be useful.  Unfortunately the current space I have been using it for I am not able to publish, so don't currently have any.  Though the general concept of what I have been doing is talked about in the paper, labeling document sections to create training/test sets for ML tools that automatically bin documents by subjects.
