### 2024-03-16 #5336 @RuneBlaze
@chrstahl Sorry for the delay (I said I would check two days ago.. Welp at least I got back in the weekend). Thanks for the README update -- I updated my checklist to reflect your personal fork. I am still leaving the issues in the USEPA/peat up because the issues reflect the status of the submitted repo -- I will happily close them after the fixes get merged.

For me at least, the only two blockers are:

 1. Merging everything into the submitted repo
 2. The only writing suggestion I had for the paper (copied from my comment above):

> One minor writing suggestion:
>
> "While text mining approaches – including Deep Learning (DL), Artificial Intelligence (AI), and Machine Learning (ML) - continue to expand at a rapid pace, the tools used by researchers to create the labeled datasets required for training, modeling, and evaluation remain rudimentary."
>
> DL is usually a specific kind of ML, so rephrasing will be appreciated.

cc: @atrisovic
