### 2024-03-18 #5336 @chrstahl
@RuneBlaze No problem.  Everything has been merged and I reworded that section and merged the MD update.  Generated a new PDF above.  I think that covers all of your blocks.

@flor14 All of your code fixes and suggestions should also be full incorporated in the repo linked to the project now also.
