### 2024-04-30 #5336 @atrisovic
Hello @flor14, I hope all is well. Just a gentle reminder about the paper [PDF Entity Annotation Tool (PEAT)](https://github.com/openjournals/joss-reviews/issues/5336#top). Thank you for your efforts!
