### 2024-05-30 #5336 @flor14
Hello

Sorry for the late reply. I did my review a while ago when I had more available time. Now it has been difficult to find time for this.

There are some inconsistencies in the citations, particularly the software citations.
In the article as it stands, in the reference section says:

> spaCy. (2022). In GitHub repository. explosion. https://github.com/explosion/spaCy

and when you enter [the repository](https://github.com/explosion/spaCy), in the section `cite this repository`, you will find:

> Honnibal, M., Montani, I., Van Landeghem, S., & Boyd, A. (2020). spaCy: Industrial-strength Natural Language Processing in Python. https://doi.org/10.5281/zenodo.1212303

This should be changed in the references section and in the article's text for all the software citations.

I think it is important to correctly credit the software developers, especially in a journal like JOSS.
