### 2024-06-04 #5336 @chrstahl
@flor14 https://github.com/USEPA/peat/pull/7 should fix the citations.  Is there anything else to finish out the review?
