### 2024-11-04 #5336 @atrisovic
Hi @flor14
I hope you are doing well. I wanted to follow up on your review of PDF Entity Annotation Tool. Your checklist is nearly complete, and we truly appreciate your time and effort on this review. If there’s anything we can assist you with to wrap things up, please let us know!
