### 2024-12-23 #5336 @atrisovic
Hi @chrstahl, can you incorporate changes regarding the citations and the license? I wonder why the open PR is not merged?
