### 2025-01-06 #5336 @chrstahl
@atrisovic I reached out to them directly, so hopefully that merge request gets merged quickly now that everyone should be back from vacation.
