### 2025-02-12 #5336 @atrisovic
Can you fix this reference

```The opening of the
58 PDF standard in 2008 (“Document Management — Portable Document Format — Part 2,”
59 2022) has helped facilitate PDF text extraction, but document formatting and content```
