### 2025-02-17 #5336 @chrstahl
@atrisovic Absolutely, I'll look over this and update with the DOI and after the reference change has been merged back.
