### 2025-03-04 #5336 @chrstahl
@atrisovic FYI reference should be fixed, I have a release version made and a pull request open on that...just waiting for additional guidance from IT to tie in Zendo to the repo and generate the DOI
