+++ .github/workflows/main.yml
@@ -17,7 +17,6 @@ jobs:
     - uses: actions/setup-node@v4
       with:
         node-version: 22
-        cache: 'yarn'
     - name: install
       run: |
         yarn
