+++ .github/workflows/main.yml
@@ -10,19 +10,20 @@ on:
   workflow_dispatch:
 
 jobs:
-  name: Build on ${{ matrix.os }}
-  runs-on: ${{ matrix.os }}
-  strategy:
-    matrix:
-      os: [windows-2022]
-  steps:
-  - uses: actions/checkout@v4
-  - uses: actions/setup-node@v4
-    with:
-      node-version: 22
-  - name: install
-    run: |
-      yarn
-  - name: build
-    run: |
-      yarn package
+  package:
+    name: Build on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [windows-2022]
+    steps:
+    - uses: actions/checkout@v4
+    - uses: actions/setup-node@v4
+      with:
+        node-version: 22
+    - name: install
+      run: |
+        yarn
+    - name: build
+      run: |
+        yarn package
