+++ .github/workflows/main.yml
@@ -11,11 +11,8 @@ on:
 
 jobs:
   package:
-    name: Build on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [windows-2022]
+    name: Windows build
+    runs-on: windows-2022
     steps:
     - uses: actions/checkout@v4
     - uses: actions/setup-node@v4
@@ -29,5 +26,6 @@ jobs:
         yarn package
     - uses: actions/upload-artifact@v4
       with:
-        path: .
+        name: peat-windows
+        path: ./release/*.exe
         retention-days: 7
