### 2025-03-09 #5336 @atrisovic
@chrstahl
Thank you!

1. Can you make sure the name of the archive on Zenodo matches the name of the paper?
2. References are not in a good shape, could you review them?
For example:

- There are a few references that have "" in the name, those should be removed (ie (“Electron,” 2022) => (OpenJS, 2022)
- some citations include full author lists, while others are abbreviated
- Date accessed is missing for all the URL references

OpenJS Foundation. (2022). Electron: A framework for building cross-platform apps. Retrieved March 9, 2025, from https://github.com/electron/electron
Mozilla. (2022). PDF.js: A web standards-based JavaScript library for parsing and rendering PDFs. Retrieved March 9, 2025, from https://github.com/mozilla/pdf.js/
International Organization for Standardization. (2022). ISO 32000-2:2020: Document management – Portable document format – Part 2. Retrieved March 9, 2025, from https://www.iso.org/standard/75839.html
