### 2025-03-18 #5336 @atrisovic
Even though I recommend acceptance, I kindly ask the @openjournals/dsais-eics to advise on the references. Please see the comments (https://github.com/openjournals/joss-reviews/issues/5336#issuecomment-2708982989), (https://github.com/openjournals/joss-reviews/issues/5336#issuecomment-2718745113) and (https://github.com/openjournals/joss-reviews/issues/5336#issuecomment-2734511896) for details.
