### 2025-03-22 #5336 @crvernon
@atrisovic  - **A note**:  we usually require checklist be completed for all reviewers.  I did not see a checklist generated by @flor14 but did see where they responded in the affirmative for completeness, so I am using this as grounds to move forward.

:wave: @chrstahl - please address the following before I move to accept this for publication:

**In the archive:**
- [ ] Please edit the metadata of the Zenodo archive to have the archive title match the paper title exactly.  No new release necessary.

**In the paper:**

- [ ] LINE 62:  missing space between "Dextr" and the citation.
- [ ] LINE 80:  "Figure 1" is not contained / referred to in the sentence.  Should this appear as, "...intended format (Figure 1)."?
- [ ] LINE 103:  No reason to capitalize the "M" and "L" in machine learning here.

Concerning the referencing issues @atrisovic noted:

I believe the following should fix your references and the in-text citations:

```bibtex
@misc{electron,
  author = {{Electron Authors}},
  title = {{Electron}},
  url = {https://github.com/electron/electron},
  year = {2025},
  note = {Accessed: 2025-03-22}
}

@misc{pdf,
  author = {{International Organization for Standardization}},
  title = {ISO 32000-2:2020 Document management -- Portable document format -- Part 2: PDF 2.0},
  url = {https://www.iso.org/standard/75839.html},
  note = {Accessed: 2025-03-22},
  year = {2020},
}

@misc{pdfjs,
  author = {Mozilla},
  year = {2025},
  title = {{PDF.js}},
  url = {https://github.com/mozilla/pdf.js},
  note = {Accessed: 2025-03-22}
}
```

Feel free to changed the accessed date as you see fit.

Let me know when you have addressed these and I will re-evaluate.  Thanks.
